move reorderer to different package
